Resolve DAT-16

See merge request Alex/dataviz!2

+++ Back/backend/migrations/20200310205650-create-temperature.js
@@ -1,36 +1,36 @@
 'use strict';
 module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.createTable('Temperatures', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER
-      },
-      value: {
-        type: Sequelize.DECIMAL
-      },
-      value_max: {
-        type: Sequelize.DECIMAL
-      },
-      value_min: {
-        type: Sequelize.DECIMAL
-      },
-      feeling: {
-        type: Sequelize.DECIMAL
-      },
-      createdAt: {
-        allowNull: false,
-        type: Sequelize.DATE
-      },
-      updatedAt: {
-        allowNull: false,
-        type: Sequelize.DATE
-      }
-    });
-  },
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.dropTable('Temperatures');
-  }
+    up: (queryInterface, Sequelize) => {
+        return queryInterface.createTable('Temperatures', {
+            id: {
+                allowNull: false,
+                autoIncrement: true,
+                primaryKey: true,
+                type: Sequelize.INTEGER
+            },
+            value: {
+                type: Sequelize.DECIMAL(10, 2)
+            },
+            value_max: {
+                type: Sequelize.DECIMAL(10, 2)
+            },
+            value_min: {
+                type: Sequelize.DECIMAL(10, 2)
+            },
+            feeling: {
+                type: Sequelize.DECIMAL(10, 2)
+            },
+            createdAt: {
+                allowNull: false,
+                type: Sequelize.DATE
+            },
+            updatedAt: {
+                allowNull: false,
+                type: Sequelize.DATE
+            }
+        });
+    },
+    down: (queryInterface, Sequelize) => {
+        return queryInterface.dropTable('Temperatures');
+    }
 };
